misc bug fixes

+++ things/models/thing.php
@@ -532,7 +532,6 @@
                              SET `oid` = '$nid' 
                            WHERE `oid` = '$pid'";
                 $sql = mysql_query ($query) or die (mysql_error ());
-                
                 return true;
             } else {
                 die ("Failed to reallocate object");
@@ -561,12 +560,29 @@
             // removes an object from the database.
             // deletes the object, the relationship with parents, and their children.
             // children of this object will become orphans.
-            $oid = $this->oid;
-            $query = "DELETE FROM `objects`
-                             WHERE `oid`='$oid'";
-            $sql = mysql_query ($query) or die (mysql_error ()); // delete object first
-            $this->DelParentsAll (); // then the properties and stuff (no orphaning on crash)
-            $this->DelPropsAll ();
+			global $user;
+			if (isset ($user) && get_class ($user) == 'User' && 
+			    class_exists ('Auth') && function_exists ('CheckAuth')) {
+				
+				if ($user->GetChildren ($this->oid) || 
+				    CheckAuth ('administrative privilege')) {
+					// ^-- CheckAuth redirects on failure by default
+					// if admin OR owner, let user delete the object
+					$oid = $this->oid;
+					$query = "DELETE FROM `objects`
+									 WHERE `oid`='$oid'";
+					$sql = mysql_query ($query) or die (mysql_error ()); // delete object first
+					$this->DelParentsAll (); // then the properties and stuff (no orphaning on crash)
+					$this->DelPropsAll ();
+				} else {
+					// not enough auth. DYING.
+					// Since CheckAuth redirects, this might never run
+					CustomException ("Cannot find enough authorisation for you to delete the object.");
+				}
+			} else {
+				// cannot find auth module. DYING.
+				CustomException ("Objects cannot be deleted unless a user is logged in.");
+			}
         }
         
     }
